+++ CHANGES.md
@@ -6,6 +6,7 @@
 
 * Issue #748: `_PARTITIONDATE` pseudo column is provided only for ingestion time **daily** partitioned tables
 * Issue #988: Read statistics are logged at TRACE level. Update the log4j configuration accordingly in order to log them.
+* Issue #990: Fix to support `allowFieldAddition` for columns with nested fields.
 
 ## 0.31.0 - 2023-06-01